Add rule to prohibit the use of the backtick shell execution operator.
Execution of shell commands should be discouraged. A number of the other commands along the same lines are discouraged via function sniffs, however, the backtick operator was so far ignored.

For this sniff to be added, the minimum PHPCS version needs to be upped as it has only just been added in the latest release.

See: https://github.com/squizlabs/PHP_CodeSniffer/pull/1073